add option to automatically follow a redirection.
If the option `{follow_redirect, true}` is given to the request, the
client will be abble to automatically follow the redirection and
retrieve the body. The maximum number of connection can be set using the
`{max_redirect, Max}` option. Default is 5.

The client will follow redirection on 301, 302 & 307 if the method is
get or head. If another method is used the tuple
`{ok, maybe_redirect, Status, Headers, Client}` will be returned. It
only follow 303 redirection (see other) if the method is a POST.

Last Location is stored in the client state in the `location` property.

ex:

    Method = get,
    URL = "http://friendpaste.com/",
    ReqHeaders = [{<<"accept-encoding">>, <<"identity">>}],
    ReqBody = <<>>,
    Options = [{follow_redirect, true}, {max_redirect, true}],
    {ok, S, H, Client} = hackney:request(Method, URL, ReqHeaders,
                                         ReqBody, Options),
    {ok, Body, Client1} = hackney:body(Client).