s/pool_size/max_connections
max_connections is a better naming also used in the dispcount pool
handler, so use it as a default. For compatibility reason we are taking
pool_size when it's present in the options.

While I'm here fix the pool size management in hackney_pool: don't store
more connections than wanted in the pool and close the old one.