fix hackney_pool:start_pool/2, hackney_pool:stop_pool/1
This change make sure that we are not restarting an already existing
pool. ALso don't stop it if it's already stopped and make sure we delete
the reference to it.

While I'm here fix the doc, start_pool don't return anymore a Pid since
it's only accesible by its name. The Pid is only used for internal
purpose now.