add support for cookies.
This change add support for cookies in hackney. There is now a `cookie`
option that can be passed to the request. It can be a sinle cookie or a
list of cookies.

To parse cookies from the response a function `hackney:cookies/1` has
been added. It return all the cookies as a list of [{Key, Value}] .