reuse the connection for {ProxyHost, ProxyPort} instead of {Host,Port}
Previous logic was wrong, we passed to `hackney_connect:connect` the
Host and Port instead of the Proxy Host and Port, so the ipv6 detection was
done on the target instead of the proxy.

Also update `hackney:maybe_proxy` so we make sure to always do the proxy
handshacke.

related to #96