fix HEAD
on head request we don't expect a body so don't force the user to read
the body to finish the request. Instead do the work immediately and
return the response status and headers immediately without a reference.

Note: it fixes the shutdown error you sometimes have when you stop the
application immmediately. Also it should prevent applications leaking
since some people were forgetting to read the body right after an head
request.