don't leaks clients on errors
When an error happened the notification wasn't send to the pool when the
default pool was used by default and not set in the options. This patch also
makes sure that all error are triggerieng a cancel event. Also make sure that
all requests are monitoreThis patch also makes sure that all error are
triggerieng a cancel event. Also make sure that all requests are monitored.d

Fix benoitc/couchbeam#120