Fix multipart/form-data
Obey RFC 7578 and 2046 on multipart boundary end

When encoding the last multipart boundary, make sure it is terminated
with CRLF as RFC 2046 states it should (section 5.1.1). This unbreaks
the cowboy test and makes it pass. This also improves hackney for
multipart upload to other sources which has no special handling of this
corner case.

* Introduce Cowboy as a test dependency.
* Provide a simple upload_resource handler for testing the Hackney
  multipart upload features.