Close #291: Fix typos for configuring exometer metrics
This updates the exometer configuration instructions in the
README to match the term that `mod_metrics` expects in order
to enable exometer metrics.

This also updates `mod_metrics` so that the module name used
when exometer metrics are enabled matches the name of the
module that is provided by the metrics library.