Handle continuation lines in HTTP headers correctly
RFC2616 says:

> HTTP/1.1 header field values can be folded onto multiple lines if the
> continuation line begins with a space or horizontal tab. All linear
> white space, including folding, has the same semantics as SP. A
> recipient MAY replace any linear white space with a single SP before
> interpreting the field value or forwarding the message downstream.
>
> LWS            = [CRLF] 1*( SP | HT )
>
> SP             = <US-ASCII SP, space (32)>
> HT             = <US-ASCII HT, horizontal-tab (9)>

Hackney was not handling this correctly: it was stripping leading spaces
and tabs from continuation lines and then parsing them incorrectly as a
new header line.

This commit fixes that: `parse_header/1` now checks if the following
line starts with a space or tab and joins it onto the current line if
so.