fix owner tracking
previously we were unlinking a process that started a request each time
a stream were closing or when we were streaming to another process the
response.

This change allows a process to start multiple requests and stream. It
fix it by doing the following:

* only streams and owners are supervised by the manager to track the
socket life
* the stream is responsible of monitoring its target. If the target
process die it exists and the manager clean it
* a process can now have multiple client referenced to it. (which is
done via the addition of track_owner/3 and untrack_owner/3 functions in
the manager).

fix #441