Fixes test run example in the README
Originally the command for running the tests was `make test`. There is no `make
test` task, but there is `rebar3 eunit` which is mentioned higher in the README.
I tested the tests replacing `make test` with `rebar3 eunit` and everything ran
so I assume this is the correct way of running the tests.

Amos King @adkron <amos@binarynoggin.com>