hackney_pool: make sure to reuse only connection with same options
pool: fix usage with proxy and ssl connections

When a proxyied connection is used the connection open to the proxy can
only be used with its initial target. Similarly when reusing an SSL
connection we should make sure we are using the same options.

This change ensure above actually happens by using a connection record
as ref that contains the connections options and tunnel target instead
of only using the initial connection tuple `{Transport, Host, Port}`.

* Changes:

- create an `hackney_connection` module to handle the connection record.
- replace the Checkinreference by a #connection{} record thta conntains
all useful informations (Host, Port, Transport, ConnectoinOptions,
Tunnel)
- change `hackney_pool:checkout` to always return a socket if none is
available in the pool. This change is needed to preverse compatibility
with older version.

* notes:

Whil ethe compatibility with older versions is preserverd it's uup to
custom pool handlers to ensure that the checking reference will preserve
the unicity of a connection.