add new url format for unix socket
Current way to pass the socket requires the user to url encode the path.
For example "/var/run/mysocket' should be encoded as
"%2Fvar%2Frun%2Fmyscoket" which is not really convenient.

Changes:

We introduce thiss tuple as a ne format {<<"unix:/var/run/mysock'>>,
<<"http://path/to/resource"> }

This allows the user to pass a path withut encoding it while supporting
extensions in the future.

note: for now only http is suppored.