ensuure we release a socket in the pool
ensure we release a socket in the pool when the requester died.

Changes:

Instead of using spawn_link we just spawn the checkout process. Once the
result is returned from the pool we check if the process died. If it
dies just return the socket in the pool and continue.